+++ R/phylospatial-methods.R
@@ -8,6 +8,7 @@ print.phylospatial <- function(x, ...){
           " -", ncol(x$comm), "lineages across", nrow(x$comm), "occupied sites",
           paste0("(", x$n_sites, " total)"), "\n",
           " - community data type:", x$data_type, "\n",
+          " - branch length rescaling:", ifelse(is.null(x$rescale), "sum1", x$rescale), "\n",
           " - spatial data class:", class(x$spatial)[1], "\n",
           " - dissimilarity data:", ifelse(is.null(x$dissim), "none", x$dissim_method), "\n")
 }